+++ docs/fr_FR/Changelog.rst
@@ -4,6 +4,7 @@ ChangeLog
 - Page maintenance: Infos clefs: Mise-à-jour.
 - Interne: Mise-à-jour 'info.json'. Version minimale du core = 4.0.
 - Dependances: Mise-à-jour du code. Unique dépendance de base = python3.
+- Zigate WIFI: Socat relancé au bout de 2 mins si pas de signe de la Zigate.
 
 230718-BETA-1
 -------------